backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kitsu-backend"
version = "0.1.0"
description = "FastAPI backend rewrite for Kitsu"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.115.5,<1.0.0",
  "uvicorn[standard]>=0.32.0,<0.33.0",
  "python-multipart>=0.0.18,<0.1.0",

  # Database
  "sqlalchemy[asyncio]>=2.0.36,<2.1.0",
  "alembic>=1.13.2,<2.0.0",
  "asyncpg>=0.29.0,<1.0.0",

  # Validation / config
  "pydantic>=2.9,<3.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "email-validator>=2.2.0,<3.0.0",

  # Auth
  "bcrypt>=4.0.0,<5.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "PyJWT>=2.9.0,<3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
